+++ react/EXPERIMENTAL_Filter.js
@@ -1,3 +1,0 @@
-import EXPERIMENTAL_Filter from './components/EXPERIMENTAL_Filter/index'
-
-export default EXPERIMENTAL_Filter