+++ react/EXPERIMENTAL_FilterBar.js
@@ -1,0 +1,3 @@
+import EXPERIMENTAL_FilterBar from './components/EXPERIMENTAL_FilterBar/index'
+
+export default EXPERIMENTAL_FilterBar