+++ styleguide.config.js
@@ -159,7 +159,7 @@ module.exports = {
           components: [
             'react/components/EXPERIMENTAL_Conditions/index.js',
             'react/components/EXPERIMENTAL_Select/index.js',
-            'react/components/EXPERIMENTAL_Filter/index.js',
+            'react/components/EXPERIMENTAL_FilterBar/index.js',
           ],
         },
       ],